Implement full CRUD functionality for 'trips' and 'stops'
Flask routing and templates created which enable the user to perform the
following tasks:
1) Trips: Add, Update, Remove
2) Stops: Add, Duplicate, Update, Remove

When a 'trip' is removed, all 'stops' associated with the trip are also
removed.

Flash messages have been added to provide user feedback throughout.

Collections have been stored in variables to streamline code.

A dummy function has been created to check for user permissions, this
will be implemented in the 'users' branch, following
the merging of the 'trips' branch to master.