Removes old user functions and updates navbar to respond to user login
This is preparatory work for implementation of full user login
functionality.

The navigation bar links have been updated to
display a set of links dependent on if the user is logged in or not.

The javascript fetch code has been removed as this will no longer be
used. The flask route was also removed from the app.py.