Adds template for updating a 'stop'
New template created which is used to enable the user to update existing
stops. This is effectively the same as the stop_new.html template with
the addition of a pass-through stop query, which provides pre-populated
values for the form inputs.

This template is also used for the 'duplicate stop' functionality. A
stop is duplicated, then the user is redirected to this template to
perform any changes to the duplicated entry.