Adds templates for registering new user and enabling user login
The templates combined with the changes to the app.py enable user
management.

Users login by inputting their username, this is checked to the database
if the user exists, a 'username' session variable is created which is
used throughout the templates to determine what the user should see.
The session variable stores the unique user id from the database.
This makes it easy to check whether the user is the owner of a trip.

User registration takes a few basic user details and checks the database
determine if this username already exists. If not, it is created.
If it already exists, the user is prompted to retry.