Adds new field for trips to capture number of travelers
A new field has been added to capture the number of travelers per trip.
This will be used to scale the trip total costs. Additional text added
to input form to clarify what the costs represent.

Trips overview page is updated to include number of travelers and the
total cost is updated to multiply by the number of travelers.