Updates detailed trip view to only use one query
One aggregate query is used to pull the required information.

The query data is processed via a loop, with two data variables created
from this.

The first contains trip level detail which is used for the 'overview'
section.

The second contains stop detail, a list containing a dictionary for
each stop.

The two data variables are passed through to the template for display.