Adds new function and updates trip listing aggregation query
getTripDuration() function added which returns the sum of the trip's
stop durations. This is used to calculate the trip end date.

Given the above, the trip collections field 'end_date' is no longer used
. It is instead populated with an empty value so that it may be used
for future updates.

The show_trips() view aggregation query has been updated to calculate
the end date for each trip, and to produce a listing of countries when
a trip is made up of more than one.

When adding or updating a trip stop, the trip end date field has been
replaced with trip projected end date. This projects the end date based
on the duration of the current stop summed with the duration of all
other stops for this specific trip.
It makes use of two hidden fields (total_trip_duration and
current_stop_duration) to calculate.