Updates the login process to transform username string to lowercase
The username string is transformed to lower case to compare to the
usernames stored in the database which are all lowercase.

This update takes place in the user_login view and the LoginForm
validation.

If this was not corrected a user would not be able to log in to their
account if they had any capitalisation in their input.