Updates the trip_detailed view for better error handling
Checks to determine if the trip_detailed query returns any results have
been updated.

Following unit testing, it was apparent that the checks to determine if
a trip exists or not were not working. The checks were taking
run too early in the process. They have been moved and updated to
correctly redirect the user when a trip does not exist.

Flash messages have had full stops added to the end for consistency.