Updates except statements and removes unnecessary permissions checks
Bare except statements updated to capture all exceptions.

Removed unnecessary checks to determine whether the user is logged in or
not. These are not required because the same functions already check if
the user owns the trip/stop - so by default will check if they are
logged in to an account.