Updates tests to add valid and invalid data flag and refactors code
The tests for inputting data now all have a flag to indicate whether the
data is valid. This valid_entry flag will determine which unit test to
perform on an entry-by-entry basis. This provides a more complete
testing framework.

Code refactored to remove unnecessary functions which are now all passed
through the submit_form function. This makes the code easier to read
and maintain.